feat(gpu): signed mul with tests and benchmarks
