+++ tfhe/src/integer/gpu/server_key/radix/tests_signed/test_mul.rs
@@ -1,0 +1,27 @@
+use crate::integer::gpu::server_key::radix::tests_unsigned::{
+    create_gpu_parametrized_test, GpuFunctionExecutor,
+};
+use crate::integer::gpu::CudaServerKey;
+use crate::integer::server_key::radix_parallel::tests_cases_signed::{
+    signed_default_mul_test, signed_unchecked_mul_test,
+};
+use crate::shortint::parameters::*;
+
+create_gpu_parametrized_test!(integer_unchecked_mul);
+create_gpu_parametrized_test!(integer_mul);
+
+fn integer_unchecked_mul<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::unchecked_mul);
+    signed_unchecked_mul_test(param, executor);
+}
+
+fn integer_mul<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::mul);
+    signed_default_mul_test(param, executor);
+}
